examples/bouquetfl/pyproject.toml
[tool.poetry]
name = "bouquetfl"
version = "0.1.0"
description = ""
authors = ["arnogeimer"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
isort = "^8.0.1"
flwr-datasets = "^0.6.0"
flwr = {extras = ["simulation"], version = "^1.26.1"}
numpy = "^2.4.2"
pandas = "^3.0.1"
torch = "^2.10.0"
pre-commit = "^4.5.1"

[project]
name = "bouquetfl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "flwr-datasets>=0.6.0",
    "flwr[simulation]>=1.26.1",
    "ipython>=9.10.0",
    "isort>=8.0.1",
    "keyring>=25.7.0",
    "keyrings-alt>=5.0.2",
    "numba>=0.64.0",
    "psutil>=7.2.2",
    "timm>=1.0.25",
    "torch==2.7.0",
    "torchvision>=0.22.0",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "arnogeimer"

[tool.flwr.app.components]
serverapp = "bouquetfl.server_app:app"
clientapp = "bouquetfl.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 3
fraction-fit = 1
fraction-evaluate = 1
local-epochs = 5
batch-size = 64                # training batch size
learning-rate = 0.01 
num-partitions = 10
